requirements.txt
django-mathfilters==1.0.0